pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "mapfbench"
version = "2.0b"
description = "A simple library for benchmarking MAPF algorithms"
readme = "README.md"
requires-python = ">= 3.9"
authors = [
    {name = "Stefano Lanza"},
    {name = "Luca Leonzio"}
]
license = {text = "GPLv3"}
dependencies = [
    "numpy",
    "psutil",
    "pyyaml",
]



[project.optional-dependencies]
build = ["build", "setuptools"]
test = ["pytest"]
docs = ["myst-parser", "sphinx", "sphinx-design"]


[tool.setuptools.packages.find]
where = ["mapfbench"]
exclude = ["test", "docs", "sketch", ".readthedocs.yaml", "requirements.txt", ".gitignore", "dist"]

[project.scripts]
mapfbench = "cli.cli_interface:mapfbench"


[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib"
]
pythonpath = "src"